test: move methods to test utils and refactored test methods accordingly
